requirements.txt
aiogram>=3.0.0
telethon>=1.34.0
python-dotenv
